bf-crawler/requirements.txt
scrapy
news-please
google-cloud-datastore
tldextract